Create layout adapters
